Adds working delete and update to the character page and sheet

+++ src/@owl-factory/cache/functionality/crud.ts
@@ -41,7 +41,11 @@ export async function createMany<T extends RefRequired>(
  * @param ref The ref of the document to delete from the database
  * @returns The deleted document or an error explaining why it could not be deleted
  */
-export async function del<T extends RefRequired>(this: CacheController<T>, ref: Ref64): Promise<ResponseDoc<T>> {
+export async function del<T extends RefRequired>(
+  this: CacheController<T>,
+  ref: Ref64 | undefined
+): Promise<ResponseDoc<T>> {
+  if (!ref) { return {}; }
   const docs = await this.deleteMany([ref]);
   return docs[0];
 }
